Making compact phrase table binary compatible across architectures

+++ moses/CompactPT/PhraseDictionaryCompact.cpp
@@ -86,17 +86,17 @@ bool PhraseDictionaryCompact::Load(const std::vector<FactorType> &input
 
   std::FILE* pFile = std::fopen(tFilePath.c_str() , "r");
   
-  size_t indexSize;
+  uint64_t indexSize;
   if(m_inMemory)
     // Load source phrase index into memory
     indexSize = m_hash.Load(pFile);
   else
     // Keep source phrase index on disk
     indexSize = m_hash.LoadIndex(pFile);
-
-  size_t coderSize = m_phraseDecoder->Load(pFile);
+    
+  uint64_t coderSize = m_phraseDecoder->Load(pFile);
   
-  size_t phraseSize;
+  uint64_t phraseSize;
   if(m_inMemory)
     // Load target phrase collections into memory
     phraseSize = m_targetPhrasesMemory.load(pFile, false);
